make PSKs and TLS certificate variables optional
Users are no longer required to specify pre-shared keys or TLS
certificates. The role takes care of using and synchronizing pre-shared
keys and certificates already present on the nodes. If a key or
certificate is not present, the role generates a new one. Users are also
able to enforce keys and certificates regeneration. The posibility to
specify custom keys and certificates still exists.