do not fail if openssl is not installed
OpenSSL is used in the role and automated tests to generate random
preshared keys. Previously, OpenSSL was a dependency of pcs. That is no
longer the case since pcs-0.10.8-2.

In the role, keys are no longer generated on the controller, so the role
does not install OpenSSL there. Instead, preshared keys are now
generated on target nodes and the role newly ensures OpenSSL is installed on
target nodes.

In tests, OpenSSL is needed on the controller to test cases when
preshared keys are distributed from the controller. Tests ensure OpenSSL
is installed on the controller.