fix roles in constraints tests
As a side-effect of resetting stonith-watchdog-timeout cluster property
when (de)configuring SBD, pcs may create an empty CIB file. To ensure
compatibility with all pacemaker versions, the CIB is set to comply with
the lowest schema version shipped with pacemaker 2.0.0. This schema
doesn't support new role names. Therefore, even with the most recent pcs
and pacemaker, legacy roles may be stored in CIB.