s/ansible_play_hosts_all/ansible_play_hosts/ where applicable
The ha_cluster role is sometimes included within other roles.
In the scenario when a play is run against multiple hosts, and some of
them fail prior to including the ha_cluster role, the ha_cluster role
must continue only on the active hosts. Otherwise, the ha_cluster role
tries to operate on failed nodes and fails.