Allow enabled SBD on disabled cluster (#81)
* Allow enabled SBD on disabled cluster

Currently the sbd.servicewill not be enabled if the cluster autostart
is disabled. This is not intended behavior as is will effectively break
the feature. We can simply remove the condition to depend on
ha_cluster_start_on_boot as on a RHEL8 system the sbd.service has a
dependencies (Before/After/PartOf/RequiredBy) to cluster related
services which make sure it is only ever started by the cluster (a
manual start is not possible).

* code review fixes

Co-authored-by: Tomas Jelinek <tojeline@redhat.com>