refactor: update qdevice for upcoming pcs release
New pcs release drops several CLI commands used by the role to configure
qdevice certificates. On the other hand, it provides access to the
features via API.

This patch adds a custom module to expose pcs API to the role. It also
adds a module to handle configuring qdevice certificates.