fix: cast sbd option value to string
Cause: Some versions of Ansible/Jinja have a problem with evaluating
a non-string value as a string inside of a macro.

Consequence: Ansible will give an error that a templated value
is not a string.

Fix: Cast the value to a string.

Result: Ansible correctly templates the variable with no errors.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>