fix: be consistent with multipele attr sets
Currently, pcs does not support multiple nvsets and system role must
deal with it. The case of multiple nvsets in node attributes was handled
differently than multiple nvsets in other cases. This fix makes dealing
with node attributes consitent with the rest.

Also in README.md is now explicitly mentioned, that (in the cases of
multiple nvsets) only the first set is used and the rest are ignored.