refactor: generalize JsonMissingKey exception
JsonMissingKey was introduced to provide aditional information for a
user when KeyError is raised.

There are more exception types than KeyError that we need to deal during
processing source data (e.g. TypeError when on a place where dict is
expected, is a list and the list is accessed by key).

So, the first step to deal with those exceptions is use more general
exception that is not focused only on missing key.