test: clean up certs/keys created by tests
We are seeing test failures on some platforms caused by leftover keys and certs.
Make sure to clean these up, and make sure the test that uses the
certificate system role starts with no certs/keys.

In addition - on el9 systems we need to set the crypto policy to be able
to use Post Quantum Crypto (PQC) in tests.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>